reformat the first 3 groups of the UUID
reformat to match the uuid shown by the dmidecode output

Issue: UUID of a virtual machine as shown by the dmidecode output
does not match as reported by the virtual-host-gatherer
for hardware versions >=13 in VMWare environment.

Example:
On a virtual machine with hardware version 13:

localhost:~ # dmidecode -s system-uuid
1b4e2242-b3f0-55bd-39c2-263f3860836f

virtual-host-gatherer reports:
42224e1b-f0b3-bd55-39c2-263f3860836f

This code change in virtual-host-gatherer fixes this issue
so we can match the system in SCC correctly.

In addition, virtual-host-gatherer will also be reporting the
uuid reported by vmware in the optionalVmData field.