Fix pylintrc to work with latest pylint versions
The init-hook script that is currently specified in the .pylintrc
file uses execfile() which is a Python 2-ism that has been dropped
in Python 3, and backwards compatibility support for it has been
dropped in the latest versions of pylint itself.

Some research shows that the exec() based pattern:

    exec(compile(open(script).read(), script, 'exec'), ...)

is equivalent to the existing execfile() pattern:

    execfile(script, ...)

and is also compatible with older and newer versions of pylint.

This change updates the init-hook script to use the new exec() based
pattern, allowing the .pylintrc script to work with older and newer
versions of pylint.

Fixes #21